Don't base64-encode embedded SVG images
I don't see any reason for this added complexity. If this was added for
space-saving reasons, the size difference is minimal and probably even
less after gzip compression (which should be the default on modern HTTP
deployments).

Case in point: The "light" CSS file was 2278 Bytes bigger after this
change, but the gzip-compressed file was 1687 Bytes smaller.